Make Concessions playable in Revolution Phase (#476)
* Add the concession backend model

* Prepare backend for revolution phase

Refactor some backend functions in preparation for adding the revolution phase.

* Allow players to reveal their concession secrets

Allow players to reveal their concession secrets (give them up) during the revolution phase. The concessions don't actually get assigned to the senators yet (critical part of the feature), and there are no action logs issued during the revolution phase.

* Enable concession generation

Enable concessions to be generated and appear on the frontend

* Fix concession name field length

Add a migration to fix the concession name field that was too short.

* Add action log for new concession

* Create senator action log relations

Make the concession helper create a senator action log relation when a New Concession action log is created. Change the way action logs are retrieved, fixing a bug in the senator History tab. Fix a bug in the game ender that caused the wrong faction to be selected as winner if there was a tie in combined influence.

* Merge migrations

Merge the two migrations into one migration to reduce the diff.

* Fix mortality phase test

Fix issue with the mortality phase test where some of the assertions expected fewer messages than are actually returned. This issue was caused by a fix to the `delete_old_actions` function that correctly communicated the deletion of old actions via websocket messages, increasing the number of actions generated.