Allow concessions to be assigned during the faction phase (#515)
* Switch to using `List` for type hints

Switch to using `List` for type hints rather than `list`, because apparently it's better.

* Move `generate_assign_concessions_action` function

Move the `generate_assign_concessions_action` function to the concession helper script because it's no longer going to be exclusive to the revolution phase.

* Fix concession term links

Fix UI issue causing some concession term links to be broken.

* Allow concession assignment in faction phase

Make the changes that allow concessions to be assigned during the faction phase. Also update the tests so that they cover this scenario.

* Fix nested anchor issue in UI

Fix a UI issue where an anchor was nested inside another anchor. This was fixed by making the outer anchor into a div, but it still behaves like an anchor.

* Remove arrows between flags in actions area

Remove arrows between Faction flags in the actions area because it's easier to not have them. Might add them back at some point.

* Fix typo

Fix typo that was preventing the game from progressing beyond turn 1.