feat: add mortality (#538)
* feat: add user-visible logging

The logs allow users to see what happens in games.

* feat: add orthographic guidelines

* feat: tweak color

* fix: fix senator attributes in the UI

* feat: emphasize user's faction

By placing emphasis on the user's faction, they can more clearly see which faction is theirs.

* feat: add mortality

* refactor: introduce concept of senator generations

When a senator dies, the next senator to appear in the same family will be of the next generation. In the UI, generations are represented by a Roman numeral after the senator's name.

* feat: improve logging and UI responsiveness/styles

* feat: make action buttons stick to bottom