feat: combat calculator (#558)
* feat: combat calculator

* feat: add multiple tabs to the combat calculator

* feat: make the combat calculator draggable

* feat: add mobile support for combat calculator

* feat: allow reset of combat calculator position

* refactor: move calc state to CombatCalculator

* fix: bug on battle selection buttons

* feat: real time inter client synchronization

* fix: jittery changes on combat calculator

* fix: failure to show unselected commander or war

* fix: latency when using slider inputs

* refactor: move selection state to game container

Moving the selection state up to the game container so that it can be accessed by the combat calculator. This step will allow the combat calculator to modify the selection, which is necessary for a "Make proposal" button in the combat calculator.

* feat: style tweak

* fix: public game state sync issue

Standardized the data structures used to send game state via WebSocket connections.

This included a fix for an issue where public game state was inconsistently structured, causing the UI to fall out of sync with the server. The regression was introduced in this PR while making the changes required for the combat calculator to sync across clients in real time.

* feat: Transfer to proposal button

Allows a combat calculator configuration to be transferred to a proposal in one click. Can only be performed by a player with the "Propose deploying forces" action available.

* fix: linting issues

* feat: see current proposal in combat calculator

* Update README.md

* feat: improve grammar in forces description